Undo some unrelated changes
